refactor(cli): use runtime schema in config CLI
- Update `buildCliConfigSchema` in `src/cli/config-cli.ts` to call `loadRuntimeConfigSchema` instead of directly cloning `buildConfigSchema().schema`
- Keep schema shape source-of-truth aligned with gateway runtime schema behavior
- Adjust `src/cli/config-cli.test.ts` by mocking `loadRuntimeConfigSchema` and asserting the expected `channels`/`plugins` schema fragments directly
